Reorganization / Restructuration of entire Project
Source code is now mainly located in the `src` directory, while
scripts (e.g., for Training) will live in a separate `scripts`
directory for now.
Renamed the uncertainty package to uncertainty_fae and integrated
the utils package into it as a preparation for adding a `setup.cfg`
and making this an installable package/application.